Ajout des corrections SonarQube

+++ mobile/app/src/main/java/site/nohan/protoprogression/Model/Chemin.java
@@ -74,7 +74,7 @@ public class Chemin {
 
 
     public static double getDistance(Point A, Point B){
-        return Math.sqrt(Math.pow(B.x-A.x,2) + Math.pow(B.y-A.y,2));
+        return Math.sqrt(Math.pow((double)B.x-A.x,2) + Math.pow((double)B.y-A.y,2));
     }
 
     @Override
@@ -107,7 +107,7 @@ public class Chemin {
     }
 
     public Point getMiddlePoint(){
-        return this.points.get((int) Math.round(this.points.size()/2));
+        return this.points.get((int) Math.round((float)this.points.size()/2));
     }
